+++ HISTORY.md
@@ -1,5 +1,19 @@
-0.5.5 (WSM)
+0.5.8 (WSM)
 * Changed the point group finding to finding the space group.
+
+0.5.7 (JJ)
+* Updated organization of unittests so that fortpy would be able to
+  compile them.
+
+0.5.6 (GLWH)
+* Added a check to make sure that the length of each "orbit" of
+kpoints divided the order of the point group. 
+
+0.5.5 (JJ)
+* Made doc strings clearer by specifying the coordinate system of the
+  k-points and clarifying that the lattice vector or grid generating
+  vectors are the columns of matrices.
+* Added a unit test for 'mapKptsIntoFirstBZ' for body-centered_cubic.
 
 0.5.4 (WSM)
 * Removed fortpy.f90 from the repository and from the makefile (since